+++ eventtrip/data_providers/__init__.py
@@ -11,12 +11,14 @@ from eventtrip.data_providers.base import MarketDataProvider
 from eventtrip.data_providers.import_provider import SnapshotImportProvider
 from eventtrip.data_providers.manual_snapshot_provider import ManualSnapshotProvider
 from eventtrip.data_providers.mock_live_provider import MockLiveProvider
+from eventtrip.data_providers.opt_in_http_provider import OptInHttpJsonProvider
 
 __all__ = [
     "MarketDataProvider",
     "ManualSnapshotProvider",
     "MockLiveProvider",
     "SnapshotImportProvider",
+    "OptInHttpJsonProvider",
     "SUPPORTED_PROVIDER_TYPES",
     "get_provider",
     "LiveProviderDisabledError",